Added files
I watched the Dev Catalyst video on how to upload the website and uploaded the code in a .txt file. This is the RIFA website that I made in hopes of winning the competition. It says in the .txt file what the css, html, and javascript is. I hope this is how I was supposed to do it. The mhtml is there just in case the txt isn't right. if neither of them are right, then I am so sorry.